refactor: split analysis/report.py into focused modules
